requirements.txt
line-bot-sdk
flask
gunicorn
pandas
requests
